fix: ensure Menus direct child are MenuItems for accessibility to work
