+++ src/components/FlagsMenu/FlagsMenu.test.tsx
@@ -1,5 +1,7 @@
 import React from 'react'
-import { render, screen } from '@testing-library/react'
+import { expect, vi } from 'vitest'
+import { ISO_CODES, MuiTelInputCountry } from '@shared/constants/countries'
+import { fireEvent, render, screen } from '@testing-library/react'
 import FlagsMenu from './FlagsMenu'
 import '@testing-library/jest-dom'
 
@@ -11,9 +13,6 @@ describe('components/FlagsMenu', () => {
   beforeAll(() => {
     document.body.innerHTML = `<div>My textfield</div>`
   })
-  test('should render correctly', () => {
-    render(<FlagsMenu isoCode="FR" onSelectCountry={() => {}} />)
-  })
 
   test('should displayed when anchorEl is valid', () => {
     render(
@@ -23,5 +22,127 @@ describe('components/FlagsMenu', () => {
         onSelectCountry={() => {}}
       />
     )
+  })
+
+  test('should render correctly', () => {
+    render(
+      <FlagsMenu
+        isoCode="FR"
+        excludedCountries={['FR']}
+        onSelectCountry={() => {}}
+      />
+    )
+  })
+
+  test('should list all countries without filters props', () => {
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        isoCode="FR"
+        onSelectCountry={() => {}}
+      />
+    )
+    expect(screen.getAllByRole('option').length).toBe(ISO_CODES.length)
+  })
+
+  test('should fire onSelectCountry', () => {
+    const callback = vi.fn((country: MuiTelInputCountry) => {
+      return country
+    })
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        isoCode="FR"
+        onSelectCountry={callback}
+      />
+    )
+    fireEvent.click(screen.getByText('France'))
+    expect(callback).toBeCalledTimes(1)
+    expect(callback).toHaveBeenLastCalledWith('FR')
+  })
+
+  test('should list onlyCountries', () => {
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        onlyCountries={['FR']}
+        isoCode="FR"
+        onSelectCountry={() => {}}
+      />
+    )
+    const options = screen.getAllByRole('option')
+    expect(options.length).toBe(1)
+    expect(options[0].id).toBe('country-FR')
+  })
+
+  test('should exclude countries', () => {
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        excludedCountries={['FR']}
+        isoCode="FR"
+        onSelectCountry={() => {}}
+      />
+    )
+    expect(screen.queryByText('France')).toBeNull()
+  })
+
+  test('should display EU countries except FR', () => {
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        excludedCountries={['FR']}
+        continents={['EU']}
+        isoCode="FR"
+        onSelectCountry={() => {}}
+      />
+    )
+    expect(screen.queryByText('France')).toBeNull()
+    expect(screen.queryByText('Venezuela')).toBeNull()
+    expect(screen.getByText('Belgium')).toBeTruthy()
+  })
+
+  test('should display onlyCountries and not continents', () => {
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        onlyCountries={['VE']}
+        continents={['EU']}
+        isoCode="FR"
+        onSelectCountry={() => {}}
+      />
+    )
+    expect(screen.getByText('Venezuela')).toBeTruthy()
+    expect(screen.getAllByRole('option').length).toBe(1)
+  })
+
+  test('should highlight preferred countries and in good order', () => {
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        preferredCountries={['FR', 'BE', 'VE']}
+        isoCode="FR"
+        onSelectCountry={() => {}}
+      />
+    )
+    const options = screen.getAllByRole('option')
+    expect(options[0]).toHaveTextContent('FR')
+    expect(options[1]).toHaveTextContent('BE')
+    expect(options[2]).toHaveTextContent('VE')
+  })
+
+  test('should highlight preferred countries not excluded', () => {
+    render(
+      <FlagsMenu
+        anchorEl={getAnchorEl()}
+        preferredCountries={['FR', 'BE', 'VE']}
+        excludedCountries={['FR']}
+        isoCode="FR"
+        onSelectCountry={() => {}}
+      />
+    )
+    const options = screen.getAllByRole('option')
+    expect(options[0]).toHaveTextContent('BE')
+    expect(options[1]).toHaveTextContent('VE')
   })
 })
